Time: 0 ms (100%), Space: 13.1 MB (59.32%) - LeetHub

+++ 3084-count-substrings-starting-and-ending-with-given-character/3084-count-substrings-starting-and-ending-with-given-character.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long countSubstrings(string s, char c) {
-        int cn = 0;
+        long long cn = 0;
         for (auto e : s) {
             if (e == c) {
                 cn++;
